+++ recipes/default.rb
@@ -1,11 +1,22 @@
 #
-# Cookbook Name:: loggly
+# Cookbook Name:: loggly_rsyslog_ng
 # Recipe:: default
 #
-# Copyright (C) 2014 Matt Veitas
+# Author: Matt Veitas mveitas@gmail.com
+# Author: Kostiantyn Lysenko gshaud@gmail.com
 #
-# All rights reserved - Do Not Redistribute
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
 #
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 
 if node['loggly']['token']['from_databag']
   databag = node['loggly']['token']['databag']
@@ -20,7 +31,7 @@ end
 
 include_recipe "rsyslog::default"
 
-include_recipe "loggly-rsyslog::tls" if node['loggly']['tls']['enabled']
+include_recipe "loggly_rsyslog_ng::tls" if node['loggly']['tls']['enabled']
 
 template node['loggly']['rsyslog']['conf'] do
   source 'rsyslog-loggly.conf.erb'