+++ spec/default_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe 'loggly-rsyslog::default' do
+describe 'loggly_rsyslog_ng::default' do
 
   before do
     Chef::EncryptedDataBagItem.stub(:load).with('loggly', 'token').and_return(