+++ test/cookbooks/loggly-rsyslog_test/metadata.rb
@@ -1,8 +1,8 @@
-name             "loggly-rsyslog_test"
-maintainer       'Matt Veitas'
-maintainer_email 'mveitas@gmail.com'
+name             "loggly_rsyslog_ng_test"
+maintainer       'Kostiantyn Lysenko'
+maintainer_email 'gshaud@gmail.com'
 license          'Apache 2.0'
 description      "Used to test the Loggly rsyslog cookbook"
 version          "1.0.0"
 
-depends "loggly-rsyslog"
+depends "loggly_rsyslog_ng"